Move InputError to match InputLabel
